requirements.txt
playsound==1.2.2
requests==2.25.1
python-telegram-bot==13.5